# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recipe Book API",
        "description": "API for managing recipes, cuisines, ingredients, and user authentication",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/cuisines": {
            "get": {
                "tags": [
                    "Cuisines"
                ],
                "summary": "Get list of cuisines",
                "description": "Get list of cuisines",
                "operationId": "6252d54f56fb0c0b70daa22afa386735",
                "responses": {
                    "200": {
                        "description": "List of cuisines",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cuisine"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cuisines"
                ],
                "summary": "Create a new cuisine",
                "description": "Create a new cuisine",
                "operationId": "9e4e8ae234b3ccb36f00539f0cfffd5e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Italian"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cuisine created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cuisine created successfully"
                                        },
                                        "cuisine": {
                                            "$ref": "#/components/schemas/Cuisine"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            }
        },
        "/api/cuisines/{id}": {
            "get": {
                "tags": [
                    "Cuisines"
                ],
                "summary": "Get a specific cuisine",
                "description": "Get a specific cuisine",
                "operationId": "6a9cc9158ee06e6bec76eb53a08805e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cuisine details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cuisine"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cuisines"
                ],
                "summary": "Update a cuisine",
                "description": "Update a cuisine",
                "operationId": "3b26ce31cd2d6d9869a98fc03ab92d70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Italian"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cuisine updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cuisine updated successfully"
                                        },
                                        "cuisine": {
                                            "$ref": "#/components/schemas/Cuisine"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cuisines"
                ],
                "summary": "Delete a cuisine",
                "description": "Delete a cuisine",
                "operationId": "646c6c1c8c8ff4f254137269d6fd22db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cuisine deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cuisine deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Cuisine is used in recipes"
                    }
                },
                "security": [
                    {
                        "session": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cuisine": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Italian"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "session": {
                "type": "apiKey",
                "description": "Session-based authentication using Laravel sessions",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}